Update redis.js
I'm unable to get this to connect to non-local Redis. I'm no NodeJS coder, but it seems from the IoRedis that the server options should be the ip and port and it looks like the get.node function would return NULL (as the n.server is not a valid node id) causing the Redis function to connect to local host always.